+++ demo/image_classifier/run.py
@@ -1,31 +1,20 @@
 import gradio as gr
 import tensorflow as tf
-import numpy as np
-import json
-from os.path import dirname, realpath, join
+import requests
 
-# Load human-readable labels for ImageNet.
-current_dir = dirname(realpath(__file__))
-with open(join(current_dir, "files/imagenet_labels.json")) as labels_file:
-    labels = json.load(labels_file)
+inception_net = tf.keras.applications.MobileNetV2() # load the model
 
-mobile_net = tf.keras.applications.MobileNetV2()
-def image_classifier(im):
-    arr = np.expand_dims(im, axis=0)
-    arr = tf.keras.applications.mobilenet.preprocess_input(arr)
-    prediction = mobile_net.predict(arr).flatten()
-    return {labels[i]: float(prediction[i]) for i in range(1000)}
+# Download human-readable labels for ImageNet.
+response = requests.get("https://git.io/JJkYN")
+labels = response.text.split("\n")
 
-iface = gr.Interface(
-    image_classifier, 
-    gr.inputs.Image(shape=(224, 224)), 
-    gr.outputs.Label(num_top_classes=3),
-    capture_session=True,
-    interpretation="default",
-    examples=[
-        ["images/cheetah1.jpg"],
-        ["images/lion.jpg"]
-    ])
+def classify_image(inp):
+  inp = inp.reshape((-1, 224, 224, 3))
+  inp = tf.keras.applications.mobilenet_v2.preprocess_input(inp)
+  prediction = inception_net.predict(inp).flatten()
+  return {labels[i]: float(prediction[i]) for i in range(1000)}
 
-if __name__ == "__main__":
-    iface.launch()
+image = gr.inputs.Image(shape=(224, 224))
+label = gr.outputs.Label(num_top_classes=3)
+
+gr.Interface(fn=classify_image, inputs=image, outputs=label).launch()
