+++ demo/image_classifier_2/run.py
@@ -1,20 +1,22 @@
-import gradio as gr
-import tensorflow as tf
+import torch
 import requests
+import gradio as gr
+from PIL import Image
+from torchvision import transforms
 
-inception_net = tf.keras.applications.MobileNetV2() # load the model
+model = torch.hub.load('pytorch/vision:v0.6.0', 'resnet18', pretrained=True).eval()
 
 # Download human-readable labels for ImageNet.
 response = requests.get("https://git.io/JJkYN")
 labels = response.text.split("\n")
 
-def classify_image(inp):
-  inp = inp.reshape((-1, 224, 224, 3))
-  inp = tf.keras.applications.mobilenet_v2.preprocess_input(inp)
-  prediction = inception_net.predict(inp).flatten()
+def predict(inp):
+  inp = Image.fromarray(inp.astype('uint8'), 'RGB')
+  inp = transforms.ToTensor()(inp).unsqueeze(0)
+  with torch.no_grad():
+    prediction = torch.nn.functional.softmax(model(inp)[0], dim=0)
   return {labels[i]: float(prediction[i]) for i in range(1000)}
 
-image = gr.inputs.Image(shape=(224, 224))
-label = gr.outputs.Label(num_top_classes=3)
-
-gr.Interface(fn=classify_image, inputs=image, outputs=label).launch()
+inputs = gr.inputs.Image()
+outputs = gr.outputs.Label(num_top_classes=3)
+gr.Interface(fn=predict, inputs=inputs, outputs=outputs).launch()
